Convert measurement values from integer to decimal.

+++ wsgi/openshift/measurement/models.py
@@ -8,7 +8,7 @@ class Measurement(models.Model):
     created_at = models.DateTimeField(default=datetime.datetime.now())
     updated_at = models.DateTimeField(auto_now=True)
     type = models.ForeignKey('Category')
-    value = models.IntegerField()
+    value = models.DecimalField(max_digits=6, decimal_places=2)
     notes = models.CharField(max_length=200, null=True)
     user = models.ForeignKey('User')
 
@@ -21,10 +21,12 @@ class Category(models.Model):
     created_at = models.DateTimeField(default=datetime.datetime.now())
     updated_at = models.DateTimeField(auto_now=True)
     name = models.CharField(max_length=50)
-    min_value = models.IntegerField(default=1)
+    min_value = models.DecimalField(max_digits=6, decimal_places=2,
+                                    default=1.0)
     min_value_desc = models.CharField(max_length=50,
                                       default='Minimally')
-    max_value = models.IntegerField(default=4)
+    max_value = models.DecimalField(max_digits=6, decimal_places=2,
+                                    default=4.0)
     max_value_desc = models.CharField(max_length=50,
                                       default='Maximally')
     hidden = models.BooleanField(default=False)
@@ -42,5 +44,5 @@ class User(MeasurementUser):
 
 class MeasurementForm(ModelForm):
     class Meta:
-       model = Measurement
-       exclude = ['user','created_at','updated_at']
+        model = Measurement
+        exclude = ['user', 'created_at', 'updated_at']
